+++ ios/RNInstabug/InstabugReactBridge.h
@@ -26,13 +26,15 @@
 
 - (void)setEnabled:(BOOL)isEnabled;
 
-- (void)init:(NSString *)token invocationEvents:(NSArray *)invocationEventsArray debugLogsLevel:(IBGSDKDebugLogsLevel)sdkDebugLogsLevel useNativeNetworkInterception:(BOOL)useNativeNetworkInterception codePushVersion:(NSString *)codePushVersion overAirVersion:(NSDictionary *)overAirVersion;
+- (void)init:(NSString *)token invocationEvents:(NSArray *)invocationEventsArray debugLogsLevel:(IBGSDKDebugLogsLevel)sdkDebugLogsLevel useNativeNetworkInterception:(BOOL)useNativeNetworkInterception codePushVersion:(NSString *)codePushVersion appVariant:(NSString *)appVariant options:(nullable NSDictionary *)options  overAirVersion:(NSDictionary *)overAirVersion;
 
 - (void)setCodePushVersion:(NSString *)version;
 
 - (void)setOverAirVersion:(NSDictionary *)overAirVersion;
 
 - (void)setUserData:(NSString *)userData;
+
+- (void)setAppVariant:(NSString *)appVariant;
 
 - (void)setTrackUserSteps:(BOOL)isEnabled;
 
@@ -43,6 +45,8 @@
 - (void)setColorTheme:(IBGColorTheme)colorTheme;
 
 - (void)setPrimaryColor:(UIColor *)color;
+
+- (void)setTheme:(NSDictionary *)themeConfig;
 
 - (void)appendTags:(NSArray *)tags;