+++ test/mocks/mockInstabug.ts
@@ -43,6 +43,7 @@ const mockInstabug: InstabugNativeModule = {
   clearAllUserAttributes: jest.fn(),
   showWelcomeMessageWithMode: jest.fn(),
   setWelcomeMessageMode: jest.fn(),
+  setAppVariant: jest.fn(),
   setFileAttachment: jest.fn(),
   addPrivateView: jest.fn(),
   removePrivateView: jest.fn(),
@@ -50,9 +51,6 @@ const mockInstabug: InstabugNativeModule = {
   setPreSendingHandler: jest.fn(),
   reportScreenChange: jest.fn(),
   reportCurrentViewChange: jest.fn(),
-  addExperiments: jest.fn(),
-  removeExperiments: jest.fn(),
-  clearAllExperiments: jest.fn(),
   networkLogIOS: jest.fn(),
   networkLogAndroid: jest.fn(),
   addFeatureFlags: jest.fn(),
@@ -78,6 +76,8 @@ const mockInstabug: InstabugNativeModule = {
   setOnFeaturesUpdatedListener: jest.fn(),
   enableAutoMasking: jest.fn(),
   getNetworkBodyMaxSize: jest.fn().mockResolvedValue(10240), // 10 KB
+  setTheme: jest.fn(),
+  setFullscreen: jest.fn(),
 };
 
 export default mockInstabug;