+++ test/modules/Instabug.spec.ts
@@ -1,7 +1,7 @@
 import '../mocks/mockInstabugUtils';
 import '../mocks/mockNetworkLogger';
 
-import { findNodeHandle, Platform, processColor } from 'react-native';
+import { findNodeHandle, Platform } from 'react-native';
 import type { NavigationContainerRefWithCurrent } from '@react-navigation/native'; // Import the hook
 import { mocked } from 'jest-mock';
 import waitForExpect from 'wait-for-expect';
@@ -70,8 +70,8 @@ describe('Instabug Module', () => {
     expect(NativeInstabug.reportScreenChange).toBeCalledWith(screenName);
   });
 
-  it("componentDidAppearListener shouldn't call the native method reportScreenChange if first screen", async () => {
-    await Instabug.init({
+  it("componentDidAppearListener shouldn't call the native method reportScreenChange if first screen", () => {
+    Instabug.init({
       token: 'some-token',
       invocationEvents: [InvocationEvent.none],
     });
@@ -82,7 +82,7 @@ describe('Instabug Module', () => {
       componentType: 'Component',
     });
 
-    await waitForExpect(() => {
+    waitForExpect(() => {
       // Only first screen should be reported
       expect(NativeInstabug.reportScreenChange).toBeCalledTimes(1);
       expect(NativeInstabug.reportScreenChange).toBeCalledWith('Initial Screen');
@@ -90,6 +90,11 @@ describe('Instabug Module', () => {
   });
 
   it("componentDidAppearListener shouldn't call the native method reportScreenChange twice if same screen", (done) => {
+    Instabug.init({
+      token: 'some-token',
+      invocationEvents: [InvocationEvent.none],
+    });
+
     Array(5).forEach(() => {
       Instabug.componentDidAppearListener({
         componentId: '1',
@@ -108,7 +113,7 @@ describe('Instabug Module', () => {
       // 2. Second+ calls:
       //    The screen name is the same as _lastScreen (stored in 1st call)
       //    so it doesn't report a screen change
-      expect(NativeInstabug.reportScreenChange).not.toBeCalled();
+      expect(NativeInstabug.reportScreenChange).toBeCalledTimes(1);
       done();
     }, 1500);
   });
@@ -284,12 +289,13 @@ describe('Instabug Module', () => {
     expect(onStateChangeMock).toHaveBeenCalledWith(mockNavigationContainerRef.getRootState());
   });
 
-  it('should call the native method init', async () => {
+  it('should call the native method init', () => {
     const instabugConfig = {
       token: 'some-token',
       invocationEvents: [InvocationEvent.floatingButton, InvocationEvent.shake],
       debugLogsLevel: LogLevel.debug,
       codePushVersion: '1.1.0',
+      ignoreAndroidSecureFlag: true,
       overAirVersion: {
         service: OverAirUpdateServices.expo,
         version: 'D0A12345-6789-4B3C-A123-4567ABCDEF01',
@@ -297,7 +303,7 @@ describe('Instabug Module', () => {
     };
     const usesNativeNetworkInterception = false;
 
-    await Instabug.init(instabugConfig);
+    Instabug.init(instabugConfig);
 
     expect(NetworkLogger.setEnabled).toBeCalledWith(true);
     expect(NativeInstabug.init).toBeCalledTimes(1);
@@ -307,6 +313,8 @@ describe('Instabug Module', () => {
       instabugConfig.debugLogsLevel,
       usesNativeNetworkInterception,
       instabugConfig.codePushVersion,
+      undefined,
+      { ignoreAndroidSecureFlag: instabugConfig.ignoreAndroidSecureFlag },
       instabugConfig.overAirVersion,
     );
   });
@@ -332,13 +340,14 @@ describe('Instabug Module', () => {
     expect(NativeInstabug.setOverAirVersion).toBeCalledWith(OTAversion);
   });
 
-  it('init should disable JavaScript interceptor when using native interception mode', async () => {
+  it('init should disable JavaScript interceptor when using native interception mode', () => {
     const instabugConfig = {
       token: 'some-token',
       invocationEvents: [InvocationEvent.floatingButton, InvocationEvent.shake],
       debugLogsLevel: LogLevel.debug,
       networkInterceptionMode: NetworkInterceptionMode.native,
       codePushVersion: '1.1.0',
+      ignoreAndroidSecureFlag: true,
       overAirVersion: {
         service: OverAirUpdateServices.expo,
         version: 'D0A12345-6789-4B3C-A123-4567ABCDEF01',
@@ -346,12 +355,10 @@ describe('Instabug Module', () => {
     };
 
     // Stubbing Network feature flags
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(true));
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(true);
     jest.spyOn(NativeNetworkLogger, 'hasAPMNetworkPlugin').mockReturnValue(Promise.resolve(true));
 
-    await Instabug.init(instabugConfig);
+    Instabug.init(instabugConfig);
 
     if (Platform.OS === 'android') {
       expect(NetworkLogger.setEnabled).not.toBeCalled();
@@ -376,6 +383,8 @@ describe('Instabug Module', () => {
         // usesNativeNetworkInterception should be true when using native interception mode with iOS
         true,
         instabugConfig.codePushVersion,
+        undefined,
+        { ignoreAndroidSecureFlag: instabugConfig.ignoreAndroidSecureFlag },
         instabugConfig.overAirVersion,
       );
     }
@@ -479,12 +488,13 @@ describe('Instabug Module', () => {
     expect(NativeInstabug.setColorTheme).toBeCalledWith(theme);
   });
 
-  it('should call the native method setPrimaryColor', () => {
+  it('should call the native method setPrimaryColor on iOS', () => {
+    Platform.OS = 'ios';
     const color = '#fff';
-    Instabug.setPrimaryColor(color);
+    Instabug.setTheme({ primaryColor: color });
 
-    expect(NativeInstabug.setPrimaryColor).toBeCalledTimes(1);
-    expect(NativeInstabug.setPrimaryColor).toBeCalledWith(processColor(color));
+    expect(NativeInstabug.setTheme).toBeCalledTimes(1);
+    expect(NativeInstabug.setTheme).toBeCalledWith({ primaryColor: color });
   });
 
   it('should call the native method appendTags', () => {
@@ -849,25 +859,6 @@ describe('Instabug Module', () => {
     expect(emitter.listenerCount(NativeEvents.PRESENDING_HANDLER)).toBe(1);
   });
 
-  it('should call native addExperiments method', () => {
-    const experiments = ['exp1', 'exp2'];
-    Instabug.addExperiments(experiments);
-    expect(NativeInstabug.addExperiments).toBeCalledTimes(1);
-    expect(NativeInstabug.addExperiments).toBeCalledWith(experiments);
-  });
-
-  it('should call native removeExperiments method', () => {
-    const experiments = ['exp1', 'exp2'];
-    Instabug.removeExperiments(experiments);
-    expect(NativeInstabug.removeExperiments).toBeCalledTimes(1);
-    expect(NativeInstabug.removeExperiments).toBeCalledWith(experiments);
-  });
-
-  it('should call native clearAllExperiments method', () => {
-    Instabug.clearAllExperiments();
-    expect(NativeInstabug.clearAllExperiments).toBeCalledTimes(1);
-  });
-
   it('should call native addFeatureFlags method', () => {
     const featureFlags: Array<FeatureFlag> = [
       {
@@ -924,7 +915,7 @@ describe('Instabug Module', () => {
     expect(NativeInstabug.willRedirectToStore).toBeCalledTimes(1);
   });
 
-  it('should register feature flag listener', async () => {
+  it('should register feature flag listener', () => {
     const callback = jest.fn();
     Instabug._registerFeatureFlagsChangeListener(callback);
 
@@ -967,10 +958,10 @@ describe('Instabug iOS initialization tests', () => {
     jest.advanceTimersByTime(1000);
   });
 
-  it('should initialize correctly with javascript interception mode', async () => {
+  it('should initialize correctly with javascript interception mode', () => {
     config.networkInterceptionMode = NetworkInterceptionMode.javascript;
 
-    await Instabug.init(config);
+    Instabug.init(config);
 
     expect(NativeNetworkLogger.isNativeInterceptionEnabled).toHaveBeenCalled();
     expect(NetworkLogger.setEnabled).toHaveBeenCalledWith(true);
@@ -980,16 +971,16 @@ describe('Instabug iOS initialization tests', () => {
       config.debugLogsLevel,
       false, // Disable native interception
       config.codePushVersion,
+      config.ignoreAndroidSecureFlag,
+      undefined,
       config.overAirVersion,
     );
   });
 
-  it('should initialize correctly with native interception mode when [isNativeInterceptionEnabled] == ture', async () => {
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(true));
+  it('should initialize correctly with native interception mode when [isNativeInterceptionEnabled] == ture', () => {
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(true);
 
-    await Instabug.init(config);
+    Instabug.init(config);
 
     expect(NativeNetworkLogger.isNativeInterceptionEnabled).toHaveBeenCalled();
     expect(NetworkLogger.setEnabled).toHaveBeenCalledWith(false);
@@ -999,16 +990,13 @@ describe('Instabug iOS initialization tests', () => {
       config.debugLogsLevel,
       true, // Enable native interception
       config.codePushVersion,
-      config.overAirVersion,
     );
   });
 
-  it('should disable native interception mode when user sets networkInterceptionMode to native and [isNativeInterceptionEnabled] == false', async () => {
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(false));
+  it('should disable native interception mode when user sets networkInterceptionMode to native and [isNativeInterceptionEnabled] == false', () => {
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(false);
 
-    await Instabug.init(config);
+    Instabug.init(config);
 
     expect(NativeNetworkLogger.isNativeInterceptionEnabled).toHaveBeenCalled();
     expect(NetworkLogger.setEnabled).toHaveBeenCalled();
@@ -1018,17 +1006,17 @@ describe('Instabug iOS initialization tests', () => {
       config.debugLogsLevel,
       false, // Disable native interception
       config.codePushVersion,
+      config.ignoreAndroidSecureFlag,
+      undefined,
       config.overAirVersion,
     );
   });
 
-  it('should display error message when user sets networkInterceptionMode to native and [isNativeInterceptionEnabled] == false', async () => {
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(false));
+  it('should display error message when user sets networkInterceptionMode to native and [isNativeInterceptionEnabled] == false', () => {
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(false);
     const logSpy = jest.spyOn(global.console, 'error');
 
-    await Instabug.init(config);
+    Instabug.init(config);
 
     expect(logSpy).toBeCalledTimes(1);
     expect(logSpy).toBeCalledWith(
@@ -1055,9 +1043,9 @@ describe('Instabug Android initialization tests', () => {
     };
   });
 
-  it('should initialize correctly with native interception enabled', async () => {
+  it('should initialize correctly with native interception enabled', () => {
     config.networkInterceptionMode = NetworkInterceptionMode.native;
-    await Instabug.init(config);
+    Instabug.init(config);
     fakeTimer(() => {
       expect(NativeInstabug.setOnFeaturesUpdatedListener).toHaveBeenCalled();
       expect(NetworkLogger.setEnabled).toHaveBeenCalledWith(true);
@@ -1067,19 +1055,19 @@ describe('Instabug Android initialization tests', () => {
         config.debugLogsLevel,
         false, // always disable native interception to insure sending network logs to core (Bugs & Crashes).
         config.codePushVersion,
+        { ignoreAndroidSecureFlag: config.ignoreAndroidSecureFlag },
+        undefined,
         config.overAirVersion,
       );
     });
   });
 
-  it('should show warning message when networkInterceptionMode == javascript and user added APM plugin', async () => {
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(true));
+  it('should show warning message when networkInterceptionMode == javascript and user added APM plugin', () => {
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(true);
     jest.spyOn(NativeNetworkLogger, 'hasAPMNetworkPlugin').mockReturnValue(Promise.resolve(true));
     const logSpy = jest.spyOn(global.console, 'warn');
 
-    await Instabug.init(config);
+    Instabug.init(config);
     fakeTimer(() => {
       expect(logSpy).toBeCalledTimes(1);
       expect(logSpy).toBeCalledWith(
@@ -1088,16 +1076,14 @@ describe('Instabug Android initialization tests', () => {
     });
   });
 
-  it('should show error message when networkInterceptionMode == native and user did not add APM plugin', async () => {
+  it('should show error message when networkInterceptionMode == native and user did not add APM plugin', () => {
     config.networkInterceptionMode = NetworkInterceptionMode.native;
 
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(true));
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(true);
     jest.spyOn(NativeNetworkLogger, 'hasAPMNetworkPlugin').mockReturnValue(Promise.resolve(false));
     const logSpy = jest.spyOn(global.console, 'error');
 
-    await Instabug.init(config);
+    Instabug.init(config);
 
     fakeTimer(() => {
       expect(logSpy).toBeCalledTimes(1);
@@ -1107,16 +1093,14 @@ describe('Instabug Android initialization tests', () => {
     });
   });
 
-  it('should show error message when networkInterceptionMode == native and user did not add APM plugin and the isNativeInterceptionEnabled is disabled', async () => {
+  it('should show error message when networkInterceptionMode == native and user did not add APM plugin and the isNativeInterceptionEnabled is disabled', () => {
     config.networkInterceptionMode = NetworkInterceptionMode.native;
 
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(false));
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(false);
     jest.spyOn(NativeNetworkLogger, 'hasAPMNetworkPlugin').mockReturnValue(Promise.resolve(false));
     const logSpy = jest.spyOn(global.console, 'error');
 
-    await Instabug.init(config);
+    Instabug.init(config);
 
     fakeTimer(() => {
       expect(logSpy).toBeCalledTimes(1);
@@ -1126,20 +1110,36 @@ describe('Instabug Android initialization tests', () => {
     });
   });
 
-  it('should show error message when networkInterceptionMode == native and the isNativeInterceptionEnabled is disabled', async () => {
+  it('should show error message when networkInterceptionMode == native and the isNativeInterceptionEnabled is disabled', () => {
     config.networkInterceptionMode = NetworkInterceptionMode.native;
-    jest
-      .spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled')
-      .mockReturnValue(Promise.resolve(false));
+    jest.spyOn(NativeNetworkLogger, 'isNativeInterceptionEnabled').mockReturnValue(false);
     jest.spyOn(NativeNetworkLogger, 'hasAPMNetworkPlugin').mockReturnValue(Promise.resolve(true));
     const logSpy = jest.spyOn(global.console, 'error');
 
-    await Instabug.init(config);
+    Instabug.init(config);
 
     fakeTimer(() => {
       expect(logSpy).toBeCalledTimes(1);
       expect(logSpy).toBeCalledWith(
         InstabugConstants.IBG_APM_TAG + InstabugConstants.NATIVE_INTERCEPTION_DISABLED_MESSAGE,
+      );
+    });
+  });
+
+  it('should initialize correctly with App variant', async () => {
+    config.appVariant = 'App Variant';
+    await Instabug.init(config);
+    fakeTimer(() => {
+      expect(NativeInstabug.setOnFeaturesUpdatedListener).toHaveBeenCalled();
+      expect(NativeInstabug.init).toHaveBeenCalledWith(
+        config.token,
+        config.invocationEvents,
+        config.debugLogsLevel,
+        true,
+        config.codePushVersion,
+        config.appVariant,
+        undefined,
+        config.overAirVersion,
       );
     });
   });